backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
pillow==9.3.0
drf-yasg==1.21.4